requirements.txt
Django==3.2.6
django-cors-headers==3.8.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
drf-nested-routers==0.93.3
Pillow==8.3.2
